create method getMoney and change name to firestore
