--- lib/app/modules/home/repositories/cash_flow_repository_interface.dart
@@ -1,5 +1,0 @@
-import 'package:fluxo_caixa/app/modules/home/models/cash_flow_model.dart';
-
-abstract class ICashFlowRepository {
-  Stream<List<CashFlowModel>> getCashFlow(String userName);
-}
